Finalizing fcm notification service
